completed form, table, graph
This commit completed all the requirements. Added the input form,
a table which the calculation can be deleted and view from, a table
which shows all the required stats, and a bar graph that shows sum.
The back end code still needs refactored and there are front end bugs
that need to be fixed.